Disable TLS specifics because JDK8 says TLSv1.3 is not a standard SSL protocol name
