--- containers/jdk-client/src/test/java/org/glassfish/tyrus/container/jdk/client/SSLFilterNoTLSDefinitionTest.java
@@ -1,31 +1,0 @@
-/*
- * Copyright (c) 2020 Oracle and/or its affiliates. All rights reserved.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0, which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * This Source Code may also be made available under the following Secondary
- * Licenses when the conditions for such availability set forth in the
- * Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
- * version 2 with the GNU Classpath Exception, which is available at
- * https://www.gnu.org/software/classpath/license.html.
- *
- * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
- */
-
-package org.glassfish.tyrus.container.jdk.client;
-
-import org.junit.Before;
-
-public class SSLFilterNoTLSDefinitionTest extends SslFilterTest {
-
-    @Before
-    public void before() {
-        System.setProperty("javax.net.ssl.keyStore", this.getClass().getResource(SERVER_KEY_STORE).getPath());
-        System.setProperty("javax.net.ssl.keyStorePassword", PASSWORD);
-        System.setProperty("javax.net.ssl.trustStore", this.getClass().getResource(SERVER_TRUST_STORE).getPath());
-        System.setProperty("javax.net.ssl.trustStorePassword", PASSWORD);
-    }
-
-}
